recomendador/requirements.txt
Flask
gunicorn
requests
pandas
numpy
scikit-learn
matplotlib
pika
python-dotenv
aiohttp